# src/main/kotlin/com/loc/auth_service/domain/model/AuthResponse.kt
package com.loc.auth_service.model.dto

data class AuthResponse(
    val accessToken: String,
    val refreshToken: String,
    val tokenType: String = "Bearer",
    val expiresIn: Long,
    val username: String,
    val role: String
) 